+++ benchmarks/src/cpp11/R/zzz.R
@@ -1,0 +1,3 @@
+.onUnload <- function(libpath) {
+  library.dynam.unload("zigrCpp11", libpath)
+}